docker_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
attributee=0.1.9=pypi_0
attrs=24.2.0=pypi_0
bidict=0.23.1=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.3.2=pypi_0
class-registry=2.1.2=pypi_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
dominate=2.9.1=pypi_0
easydict=1.13=pypi_0
expat=2.6.3=h6a678d5_0
filelock=3.16.1=pypi_0
fonttools=4.54.1=pypi_0
fsspec=2024.9.0=pypi_0
future=1.0.0=pypi_0
gaft=0.5.7=pypi_0
geos=3.8.0=he6710b0_0
huggingface-hub=0.25.1=pypi_0
hyperopt=0.2.7=pypi_0
idna=3.10=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=pypi_0
jpeg4py=0.1.4=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.7=pypi_0
lazy-object-proxy=1.10.0=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
llvmlite=0.43.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.10=py312h5eee18b_0
mkl_random=1.2.7=py312h526ad5a_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.3=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.6.68=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.15=h5eee18b_0
ordered-set=4.1.0=pypi_0
packaging=24.1=pypi_0
phx-class-registry=5.0.0=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py312h06a4308_0
protobuf=5.28.2=pypi_0
py4j=0.10.9.7=pypi_0
pycparser=2.22=pypi_0
pylatex=1.4.2=pypi_0
pyparsing=3.1.4=pypi_0
python=3.12.5=h5148396_1
python-dateutil=2.9.0.post0=pypi_0
pyyaml=6.0.1=py312h5eee18b_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
rpds-py=0.20.0=pypi_0
safetensors=0.4.5=pypi_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=py312h06a4308_0
shapely=2.0.5=py312h1789038_0
six=1.16.0=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.3=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboardx=2.6.2.2=pypi_0
timm=1.0.9=pypi_0
tk=8.6.14=h39e8969_0
torch=2.4.1=pypi_0
torchvision=0.19.1=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.5=pypi_0
triton=3.0.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.3=pypi_0
visdom=0.2.4=pypi_0
vot-toolkit=0.7.1=pypi_0
vot-trax=4.0.2=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.44.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_1
